bot/requirements.txt
python-telegram-bot==21.9
groq==0.13.0
gspread==6.1.2
google-auth==2.29.0
python-dotenv==1.0.1
